Cleanup and move more to new setup
