--- tests/_item_model.py
@@ -1,5 +1,0 @@
-"""Backward-compatible shim. See ``tests.pytest_beets_plugin._item_model``."""
-
-from .pytest_beets_plugin._item_model import MediaMeta
-
-__all__ = ["MediaMeta"]
