--- typehints/beets/test/_common.pyi
@@ -1,31 +1,0 @@
-class InputError(Exception):
-    output: str | None
-    def __init__(self, output: str | None = ...) -> None: ...
-
-class DummyOut:
-    encoding: str
-    buf: list[str]
-    def __init__(self) -> None: ...
-    def write(self, buf_item: str) -> None: ...
-    def get(self) -> str: ...
-    def flush(self) -> None: ...
-    def clear(self) -> None: ...
-
-class DummyIn:
-    encoding: str
-    buf: list[str]
-    reads: int
-    out: DummyOut | None
-    def __init__(self, out: DummyOut | None = ...) -> None: ...
-    def add(self, buf_item: str) -> None: ...
-    def readline(self) -> str: ...
-
-class DummyIO:
-    stdout: DummyOut
-    stdin: DummyIn
-    def __init__(self) -> None: ...
-    def addinput(self, inputs: str) -> None: ...
-    def getoutput(self) -> str: ...
-    def readcount(self) -> int: ...
-    def install(self) -> None: ...
-    def restore(self) -> None: ...
